requirements.txt
# FIXED VERSIONS - Compatible with Python 3.11+ and Streamlit
streamlit==1.28.1
yfinance==0.2.32
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
plotly==5.18.0
ta==0.10.2
python-dateutil==2.8.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
textblob==0.17.1
nltk==3.8.1
cryptography==41.0.7
pydantic==2.5.0
PyJWT==2.8.1
pytest==7.4.3
pytest-cov==4.1.0
tqdm==4.66.1
arrow==1.3.0
joblib==1.3.2
optuna==3.14.0
seaborn==0.13.0
matplotlib==3.8.2
sentry-sdk==1.38.0
